backend/requirements.txt
openai-whisper
setuptools-rust
deepl
openai
pathlib
os


ffmpeg (package manager install)